# server/swagger.yaml
openapi: 3.0.3
info:
  title: DLT TalentHub API
  description: |
    Backend API for DLT TalentHub - connecting talents with job opportunities in the DLT Africa ecosystem.
    
    ## Overview
    This API provides endpoints for user authentication, talent management, job postings, recruitment, and administrative functions.
    
    ## Authentication
    Most endpoints require authentication using JWT Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per 15 minutes per IP address.
    
    ## Error Responses
    All error responses follow a consistent format:
    ```json
    {
      "status": "error",
      "message": "Error description",
      "errors": ["Detailed error messages"]
    }
    ```
  version: 1.0.0
  contact:
    name: DLT TalentHub Team
    email: support@dlt-talenthub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.dlt-talenthub.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/recruiters/dashboard:
    get:
      summary: Get recruiter dashboard data
      description: Get dashboard statistics and data for the current recruiter
      tags: [Recruiters]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/RecruiterDashboard'
        '401':
          description: Not authenticated or not a recruiter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/recruiters/jobs:
    get:
      summary: Get jobs posted by the recruiter
      description: Get all jobs posted by the current recruiter
      tags: [Recruiters]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, draft, paused, closed]
          description: Filter by job status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: List of recruiter's jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      jobs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/recruiters/applications:
    get:
      summary: Get all applications for recruiter's jobs
      description: Get all job applications for jobs posted by the current recruiter
      tags: [Recruiters]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, reviewed, shortlisted, interviewed, offered, accepted, rejected]
          description: Filter by application status
        - in: query
          name: jobId
          schema:
            type: string
          description: Filter by specific job ID
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      applications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Application'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/recruiters/stats:
    get:
      summary: Get recruiter statistics
      description: Get statistics and metrics for the current recruiter
      tags: [Recruiters]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recruiter statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/RecruiterStats'
  /api/recruiters/search-talents:
    get:
      summary: Search talents for a specific job
      description: Search for talents that match the requirements of a specific job
      tags: [Recruiters]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: jobId
          schema:
            type: string
          description: Job ID to find matching talents for
        - in: query
          name: skills
          schema:
            type: string
          description: Required skills (comma-separated)
        - in: query
          name: experience
          schema:
            type: number
            minimum: 0
          description: Minimum years of experience
        - in: query
          name: location
          schema:
            type: string
          description: Preferred location
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: Matching talents
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      talents:
                        type: array
                        items:
                          $ref: '#/components/schemas/Talent'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/jobs:
    post:
      summary: Create a new job posting
      description: Create a new job posting (recruiters and admins only)
      tags: [Jobs]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Job'
        '401':
          description: Not authenticated or not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get all jobs
      description: Get a paginated list of all active jobs
      tags: [Jobs]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
        - in: query
          name: status
          schema:
            type: string
            enum: [active, draft, paused, closed]
          description: Filter by job status
        - in: query
          name: type
          schema:
            type: string
            enum: [full-time, part-time, contract, freelance, internship]
          description: Filter by job type
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      jobs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
...
  /api/admin/notifications:
    get:
      summary: Get system notifications
      description: Get system-wide notifications and alerts
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
          description: Filter by notification type
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: System notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/notifications:
    get:
      summary: Get user notifications
      description: Get notifications for the current user
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
          description: Filter by notification type
        - in: query
          name: isRead
          schema:
            type: boolean
          description: Filter by read status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/notifications/unread-count:
    get:
      summary: Get unread notification count
      description: Get the count of unread notifications for the current user
      tags: [Notifications]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread notification count
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      count:
                        type: integer
                        description: Number of unread notifications
  /api/notifications/{notificationId}/read:
    put:
      summary: Mark notification as read
      description: Mark a specific notification as read
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: string
          description: Notification ID
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Notification'
  /api/notifications/mark-all-read:
    put:
      summary: Mark all notifications as read
      description: Mark all notifications for the current user as read
      tags: [Notifications]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: All notifications marked as read
  /api/notifications/{notificationId}:
    delete:
      summary: Delete notification
      description: Delete a specific notification
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: string
          description: Notification ID
      responses:
        '200':
          description: Notification deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Notification deleted successfully
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          description: Error message
        errors:
          type: array
          items:
            type: string
          description: Detailed error messages
    Talent:
      type: object
      properties:
        _id:
          type: string
          description: Talent ID
        userId:
          type: string
          description: Associated user ID
        skills:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: Skill name
              level:
                type: string
                enum: [beginner, intermediate, advanced, expert]
                description: Skill proficiency level
              yearsOfExperience:
                type: number
                minimum: 0
                description: Years of experience with this skill
        experience:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                description: Job title
              company:
                type: string
                description: Company name
              startDate:
                type: string
                format: date
                description: Start date
              endDate:
                type: string
                format: date
                description: End date (null if current)
              description:
                type: string
                description: Job description
        education:
          type: array
          items:
            type: object
            properties:
              degree:
                type: string
                description: Degree name
              institution:
                type: string
                description: Institution name
              field:
                type: string
                description: Field of study
              graduationYear:
                type: integer
                description: Graduation year
        availability:
          type: object
          properties:
            status:
              type: string
              enum: [available, busy, unavailable]
              description: Availability status
            startDate:
              type: string
              format: date
              description: Available start date
            endDate:
              type: string
              format: date
              description: Available end date
            hoursPerWeek:
              type: number
              minimum: 0
              maximum: 168
              description: Hours available per week
        salary:
          type: object
          properties:
            min:
              type: number
              minimum: 0
              description: Minimum expected salary
            max:
              type: number
              minimum: 0
              description: Maximum expected salary
            currency:
              type: string
              default: USD
              description: Salary currency
        cv:
          type: string
          description: URL to CV document
        profileViews:
          type: integer
          minimum: 0
          description: Number of profile views
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Job:
      type: object
      properties:
        _id:
          type: string
          description: Job ID
        title:
          type: string
          description: Job title
        description:
          type: string
          description: Job description
        company:
          type: object
          properties:
            name:
              type: string
              description: Company name
            logo:
              type: string
              description: Company logo URL
            website:
              type: string
              format: uri
              description: Company website
        type:
          type: string
          enum: [full-time, part-time, contract, freelance, internship]
          description: Job type
        category:
          type: string
          description: Job category
        skills:
          type: array
          items:
            type: string
          description: Required skills
        requirements:
          type: array
          items:
            type: string
          description: Job requirements
        location:
          type: object
          properties:
            city:
              type: string
              description: City
            country:
              type: string
              description: Country
            remote:
              type: boolean
              description: Whether remote work is allowed
        salary:
          type: object
          properties:
            min:
              type: number
              minimum: 0
              description: Minimum salary
            max:
              type: number
              minimum: 0
              description: Maximum salary
            currency:
              type: string
              default: USD
              description: Salary currency
        status:
          type: string
          enum: [active, draft, paused, closed]
          description: Job status
        isUrgent:
          type: boolean
          description: Whether job is urgent
        isFeatured:
          type: boolean
          description: Whether job is featured
        recruiterId:
          type: string
          description: ID of the recruiter who posted the job
        views:
          type: integer
          minimum: 0
          description: Number of job views
        applications:
          type: integer
          minimum: 0
          description: Number of applications
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JobCreateRequest:
      type: object
      required:
        - title
        - description
        - company
        - type
        - category
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
          description: Job title
        description:
          type: string
          minLength: 50
          maxLength: 5000
          description: Job description
        company:
          type: object
          required:
            - name
          properties:
            name:
              type: string
              description: Company name
            logo:
              type: string
              format: uri
              description: Company logo URL
            website:
              type: string
              format: uri
              description: Company website
        type:
          type: string
          enum: [full-time, part-time, contract, freelance, internship]
          description: Job type
        category:
          type: string
          description: Job category
        skills:
          type: array
          items:
            type: string
          description: Required skills
        requirements:
          type: array
          items:
            type: string
          description: Job requirements
        location:
          type: object
          properties:
            city:
              type: string
              description: City
            country:
              type: string
              description: Country
            remote:
              type: boolean
              description: Whether remote work is allowed
        salary:
          type: object
          properties:
            min:
              type: number
              minimum: 0
              description: Minimum salary
            max:
              type: number
              minimum: 0
              description: Maximum salary
            currency:
              type: string
              default: USD
              description: Salary currency
        isUrgent:
          type: boolean
          description: Whether job is urgent
        isFeatured:
          type: boolean
          description: Whether job is featured
...
    Application:
      type: object
      properties:
        _id:
          type: string
          description: Application ID
        jobId:
          type: string
          description: Job ID
        talentId:
          type: string
          description: Talent ID
        status:
          type: string
          enum: [pending, reviewed, shortlisted, interviewed, offered, accepted, rejected]
          description: Application status
        coverLetter:
          type: string
          description: Cover letter
        expectedSalary:
          type: object
          properties:
            amount:
              type: number
              minimum: 0
              description: Expected salary amount
            currency:
              type: string
              description: Salary currency
        availability:
          type: object
          properties:
            startDate:
              type: string
              format: date
              description: Available start date
            hoursPerWeek:
              type: number
              minimum: 0
              maximum: 168
              description: Hours available per week
        notes:
          type: string
          description: Recruiter notes
        appliedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RecruiterDashboard:
      type: object
      properties:
        totalJobs:
          type: integer
          minimum: 0
          description: Total jobs posted
        activeJobs:
          type: integer
          minimum: 0
          description: Active jobs
        totalApplications:
          type: integer
          minimum: 0
          description: Total applications received
        pendingApplications:
          type: integer
          minimum: 0
          description: Pending applications
        recentJobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
          description: Recent job postings
        recentApplications:
          type: array
          items:
            $ref: '#/components/schemas/Application'
          description: Recent applications
    RecruiterStats:
      type: object
      properties:
        totalJobs:
          type: integer
          minimum: 0
          description: Total jobs posted
        activeJobs:
          type: integer
          minimum: 0
          description: Active jobs
        totalApplications:
          type: integer
          minimum: 0
          description: Total applications received
        applicationsByStatus:
          type: object
          properties:
            pending:
              type: integer
              minimum: 0
            reviewed:
              type: integer
              minimum: 0
            shortlisted:
              type: integer
              minimum: 0
            interviewed:
              type: integer
              minimum: 0
            offered:
              type: integer
              minimum: 0
            accepted:
              type: integer
              minimum: 0
            rejected:
              type: integer
              minimum: 0
        averageResponseTime:
          type: number
          minimum: 0
          description: Average response time in hours
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          description: Number of items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        pages:
          type: integer
          minimum: 0
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there is a next page
        hasPrev:
          type: boolean
          description: Whether there is a previous page
...